+++ front/src/modules/pipeline-processes/components/CompanyBoardCard.tsx
