+++ front/src/modules/pipeline-processes/components/EntityBoardActionBar.tsx
