+++ front/src/modules/pipeline-processes/components/NewCompanyBoardCard.tsx
