+++ front/src/modules/pipeline-processes/components/__stories__/Board.stories.tsx
