+++ front/src/modules/pipeline-processes/components/__stories__/CompanyBoardCard.stories.tsx
