+++ front/src/modules/pipeline-processes/components/__stories__/mock-data.ts
