+++ front/src/modules/pipeline-processes/states/boardColumnsState.ts
