+++ front/src/modules/pipeline-processes/states/boardItemsState.ts
