+++ front/src/modules/pipeline-processes/states/selectedBoardItemsState.ts
